Updated Comp Teleop, Variables, and started work on Auto

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/tuning/variables_and_subsystemClasses/Outtake.java
@@ -24,6 +24,7 @@ public class Outtake {
 
     }
 
+
     // Actions are below
 
     public class outtakeOut implements Action  {
@@ -75,7 +76,7 @@ public class Outtake {
 
         }
     }
-    public Action outtakeToDist(int dist) {
+    public Action outtakeToDist(double dist) {
         return new outtakeToDist(dist);
     }
